+++ T-Strore.Data/DTO/TransactionDTO.cs
@@ -3,8 +3,8 @@ namespace T_Strore.Data;
 
 public class TransactionDto
 {
-    public int Id { get; set; }
-    public int AccountId { get; set; }
+    public long Id { get; set; }
+    public long AccountId { get; set; }
     public DateTime Date { get; set; }
     public TransactionType TransactionType { get; set; }
     public decimal Amount { get; set; }
